refactor: 게임을 컨트롤러로 변경
게임이 맡은 역할이 도메인보다는 컨트롤러의 역할에 부합하여 변경
