+++ src/main/kotlin/Application.kt
@@ -1,7 +1,7 @@
-import omok.controller.Controller
+import omok.controller.Game
 import omok.view.InputView
 import omok.view.OutputView
 
 fun main() {
-    Controller(InputView(), OutputView()).start()
+    Game(InputView(), OutputView()).start()
 }
