+++ src/main/kotlin/omok/domain/board/Board.kt
@@ -1,15 +1,14 @@
 package omok.domain.board
 
-import omok.domain.judgment.BlackReferee
+import omok.domain.judgment.PlacementReferee
 import omok.domain.player.Stone
 import omok.model.toPresentation
 
 class Board(private val _positions: MutableMap<Position, Stone?> = POSITIONS.associateWith { null }.toMutableMap()) {
-    private val referee = BlackReferee()
     val positions: Map<Position, Stone?>
         get() = _positions.toMap()
 
-    fun placeStone(position: Position, stone: Stone) {
+    fun placeStone(position: Position, stone: Stone, referee: PlacementReferee) {
         require(isEmpty(position)) {
             "[ERROR] ${position.column.toPresentation()}${position.row.toPresentation()}은/는 비어있지 않습니다."
         }
